Enable or disable AOA depends on specific MQ signal
The meaning of this changes is to add ability to enable or disable
AOA library according to the specific signals sent by AppLink over mqueue.

The main idea is to prevent simultaneous using of AOA library be several
components (int our case ProjectionManager and SDL). As SDL has a lowest priority
it should release AOA as far as it receives signal from system.

HandledSiganls:
    TAKE_AOA - allow SDL to init AOA library.
    AOA_TAKEN - the answer from SDL in case AOA has been succesfully initialized.
    AOA_INACCESSIBLE- the answer from SDL in case AOA can't be innitialized.
    RELEASE_AOA - the command from system that forces SDL to shutdown AOA.
    AOA_RELEASED - the answer of the SDL after AOA has been shutdown.

Implements: CRQ APPLINK-14089
Related-issues: APPLINK-14091