APPLINK-15185_Error_in_resumption_data_db.cc
Changed method Init(). Using kInsertInitData from Prepare(kInsertInitData) & Exec()
changed to Exec(kInsertInitData) as kInsertInitData has multiple SQL queries
We have 2 overloaded Exec() methods in our code
1) One without parameters - is wrapper for sqlite3_step() and must be used
only with Prepare(const char*) which is wrapper for sqlite3_prepare(..)
2) Second with parameter - is wrapper for sqlite3_exec ( which is internal sqlite wrapper
for sqlite3_prepare, sqlite3_step and and sqlite3_finalize )
According sqlite API implementation : if we have multiple queries
divided by semicolons - sqlite3_step return data for only
first query and moved B-tree cursor to the beginning of next result row
in set.
So in case multiple queries we have to use sqlite3_exec which is
correctly working with multiple queries. And this is exactly our wrapper
Exec(const char*) .
And as we use it we don't need Prepare() as it invoked in sqlite3_exec.