Fix incorrect splitting of several JSON messages concatenated together
The problem was in CMessageBroker::onMessageReceived() - the boundary
of the next message is not found correctly, because it is calculated
by subtracting PARSED message length from the whole length, but parsed
message can be different from the original (if it contains floating
point numbers for example). The fix is to take the length of the
original message.
Also fixed error handling. Before it was causing an assert.

Fixes: APPLINK-15236