Fix Logger's TelnetApplender socket reopen on fast restart
When SDL is stopped, TelnetApplender's server socket listening on port
6676 will remain in TIME_WAIT state (if there was a client connected
to it) fr some time, depending on the operating system settings.
With the old code, if we restart SDL soon after stopping it,
TelnetApplender will fail to bind to its port, because 'it is in use'.
The fix is to set SO_REUSEADDR option to server socket when opening
it - this will allow the port to be reused when it is in TIME_WAIT
state. Note that the client port must be different from the previous
client port when the client reconnects, otherwise the connection should
fail, according to the documentation.

Fixes: APPLINK-15273