Fix race condition in Singleton and Application Manager failure to stop
The problem can be reproduced by starting SDL, but NOT starting the HMI
(chromium), then trying to register a mobile app. In this situation SDL
can't be stopped with ctrl+C.
The reason for the problem is that we can't delete the AM, because the
RequestController thread of AM is still running (waiting in a cycle for
the HMI to respond). Also a second AM is created, because in the
Singleton we set to 0 the instance pointer before deleting it and
someone calls instance() before destroy() finishes (because there is no
common lock).
The fix is to use a single mutex for the Singleton methods, introduce a
is_stopping_ flag in AM, set it to true in AM's Stop() method, and
also destroy RequestController's thread pool there.
Then check stop flag in RegisterAppInterfaceRequest::Run() and exit the
HMI waiting cycle there.

Fixes: APPLINK-8324