Fix wrong file path when AppStorageFolder is changed
When the customer is PASA absolute file path is used, in the other
cases we can have relative path set, thus we use GetAbsoluteFilePath(..).
This function is wraper above the standard realpath(..) (POSIX.1-2008).
By calling this function it doesn`t matter do we use relative or abs.
path as input argument, it always return the abs. path (if exist).

Closes-bug: APPLINK-12066