Fix deadlocks in AM when stopping SDL and log socket endless blocking
When we stop AM we unregister apps and call StopSavePersistentDataTimer().
If the timer callback by chance is active at this moment we can have a
deadlock between AM and ResumeController. To prevent it limit the lifetime
of ApplicationListAccessor and allow ResumeController to finish his job.
Another option is to move resume_controller().OnSuspend() call before
unregitering apps.
Another deadlock problem arises when while destroying AM an app tries to
register (3 seconds before that actually). Then AM tries to destroy
ResumeController, but RC is in ApplicationResumptiOnTimer() callback
and tries to get AM instance, and blocks on AM singleton's lock. The
fix is to StopRestoreHmiLevelTimer() before destroyng AM.
A third reason for SDL not stopping normally is when a logger is writing
on a socket, but the client on the other side is not reading (and is not
closing the socket neither). Then write() blocks and logger, and therefore
SDL, cannot be stopped. The fix is to put timeout option to the socket.

Fixes: APPLINK-18116