Remove unregistering app from HMI resumption queue
When an application with saved resume data is registering, we add it
to HMI state resumption timer queue in order to resume its HMI level
after 3 seconds. If during these 3 seconds the application voluntarily
unregisters, we delete its saved data. Then the app may register again
(still within these 3 seconds), without resumption this time. But the
scheduled HMI state resumption is still pending. So when the HMI
resumption timer callback is executed, it will not find application's
resumption data and an assertion will fail leading to core dump.
The fix is to remove the app id from HMI resumption timer queue when
the application is unregistering.

Fixes: APPLINK-14678