Fix_malformed_frequency_trigger_not_activated
The problem occure because when we have sequence of two or more
malformed packets. After the first malformed packet the
"last_portion_of_data_was_malformed_" is set to true till the last
packet which is the problem, because "++malformed_occurrence" is never
called. i.e we have "malformed_occurs" = 0 which prevent
"TrackMalformedMessage()" in protocol_handler_impl.cc to be called.
The main reason for all this is the "continue" call in "CreateFrame()".

Closes-Bug: APPLINK-19207