Fix_PerformInteraction(KEYBOARD)_handling_in_OnKeyboardInput
Currently when we receive OnKeyboardInputNotification from HMI we check
is there any active perform interaction. The layout is never checked
or stored in any way. Thats the reason when we receive ICON_ONLY
InteractionLayout in LimitedState to not be able to make the difference.
In order to fix that, a variable is added that store the layout type of
perform interaction rpc and when we receive OnKeyboardInputNotification
we check this variable.

Closes-Bug: APPLINK-19279