Increase Slider generic command timeout to receive HMI response
When we change Slider's position in HMI its generic command timeout is
reset to default. Until now the default timeout was equal to Slider's
own timeout. But this is not enough, because in HMI the timeout starts
and expires a little bit later than the generic timeout in SDL. When
generic timeout in SDL expires we return GENERIC_ERROR and ignore HMI
response, which comes a little bit (usually milliseconds) later.
The fix is to add some extra time to the generic timeout, above the
Slider timeout, to compensate HMI communication and processing delay.

Fixes: APPLINK-20719