Move media pipe creation before StartStream command
Before this change, streaming pipe was not created when we start
streamning service and send pipe URI to HMI in StartStream, but only
when the streaming begins. So the HMI may try to open a non-existing
pipe when it receives pipe URI.
To fix this moved pipe creation from PipeStreamer::Connect() to
PipeStreamer::PipeStreamer() and pipe unlink from
PipeStreamer::Disconnect() to PipeStreamer::~PipeStreamer().
Thus the pipes are created when SDL starts and media manager is
initialized.

Fixes: APPLINK-20156