Fixes resumption flow for Bluetooth transport during suspend.
There are two issues:
- wrong flow of resume data saving in case of unexpected disconnect over
  Bluetooth transport due to SUSPEND signal (SDL issue)
- system time changes (std::time(NULL)) after SUSPEND (AppLink/system
  issue)

Because of that, time stamp of apps in app_info.dat is extremely
different from ignition off time. With that fix resumption data for all
apps will be saved at almost same time as ignition off time, so there
will be no big difference in time as before. However issue with std::time
will be discussed with PASA since it also can lead to same issue in future.

Closes-bug: APPLINK-21859