Fix_Encryption_for_multiframe_messages
The root cuase of the problem was that the data in the FRAME_TYPE_FIRST
packet was encrypted when is received from spt. In order to be read
correctly it should be decrypted first, but instead it was read in
encrypted format, whicj caused wrong parameters to be read and wrong
area of data to be decrypted. This was the reason for the error in the
decryption.

* multiframe data is decrypted first and then totalDataBytes value
is set, and area of memory is allocated.

Related Issue: APPLINK-13960