Fix_SDL_status_change_on_invalid_certificate
The reason for the occurence of this issue is that when
set_crypto_manager() is called in StartComponents().
The certificate expiration_time_ is set in CryptoManager Init(),
which is called afterwards. The second Reason is that, when
IsCertificateUpdateRequired() is called result of mktime() is
never checked and if we have -1 returned, then the difftime result
will be with incorrect value..

- mktime() returned value is checked..
- IsCertificateUpdaterequired() is removed from set_crypto_manager().

Related Issue: APPLINK-16976