Add ProtocolManager class for protocols handling.
iOS SDK supposed to work using  so called `Hub strategy`.
To support this SDL has to maintain the EAP protocols collection
and handle their state: whether it is in use or free.

In the commit there is simple implementation for ProtocolManager class
which is responsible for protocols collection and their state.

Helper class ProtocolDescriptors describes the protocol itself. It
contains index, name and state of the certain protocol.